CHECK: use 0.30000000000000004.com instead of MS link
I prefer the above website because it explains the root causes of
floating point precision issues in mathematical terms, as well as
includes example code in many programming languages.

The MS page is more vague in its explanation *and* uses a C macro
instead of a function to check floating point approximate
equality, which is bad practice.